Made progressbar into component
